refactor: move xiaoe utility functions to lib directory
The xiaoe utility functions were moved from the utils directory to the lib directory to better organize the codebase and align with the project's structure. This change improves maintainability and consistency in the code organization.
